[VCDA-3721] Support ranges not identical to gateway ranges
