+++ pkg/cpisdk/cpi_gateway_manager.go
@@ -1,3 +1,8 @@
+/*
+   Copyright 2021 VMware, Inc.
+   SPDX-License-Identifier: Apache-2.0
+*/
+
 package cpisdk
 
 import (
