+++ pkg/vcdsdk/gateway_unit_test.go
@@ -13,85 +13,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestGetUnusedIPAddressInRange(t *testing.T) {
-
-	type TestCase struct {
-		StartIPAddress   string
-		EndIPAddress     string
-		UsedIPAddressMap map[string]bool
-		FreeIP           string
-		ErrorComment     string
-	}
-
-	testCaseList := []TestCase{
-		{
-			StartIPAddress: "1.2.3.4",
-			EndIPAddress:   "1.2.3.40",
-			UsedIPAddressMap: map[string]bool{
-				"1.2.3.4": true,
-				"1.2.3.5": true,
-			},
-			FreeIP:       "1.2.3.6",
-			ErrorComment: "NormalTest: The first IP in range should be returned",
-		},
-		{
-			StartIPAddress: "1.2.3.4",
-			EndIPAddress:   "1.2.3.40",
-			UsedIPAddressMap: map[string]bool{
-				"1.2.3.24": true,
-				"1.2.3.25": true,
-			},
-			FreeIP:       "1.2.3.4",
-			ErrorComment: "TestEdgeCaseFirstIP: The first IP in range should be returned",
-		},
-		{
-			StartIPAddress: "1.2.3.4",
-			EndIPAddress:   "1.2.3.6",
-			UsedIPAddressMap: map[string]bool{
-				"1.2.3.4": true,
-				"1.2.3.5": true,
-			},
-			FreeIP:       "1.2.3.6",
-			ErrorComment: "TestEdgeCaseLastIP: The last IP in range should be returned",
-		},
-		{
-			StartIPAddress:   "1.2.3.4",
-			EndIPAddress:     "1.2.3.4",
-			UsedIPAddressMap: nil,
-			FreeIP:           "1.2.3.4",
-			ErrorComment:     "OnlyOneIPTest: Exactly one IP is available and unused",
-		},
-		{
-			StartIPAddress: "1.2.3.4",
-			EndIPAddress:   "1.2.3.6",
-			UsedIPAddressMap: map[string]bool{
-				"1.2.3.4": true,
-				"1.2.3.5": true,
-				"1.2.3.6": true,
-			},
-			FreeIP:       "",
-			ErrorComment: "TestNoAvailableIP: There are no IPs available",
-		},
-		{
-			StartIPAddress: "1.2.3.4",
-			EndIPAddress:   "1.2.3.4",
-			UsedIPAddressMap: map[string]bool{
-				"1.2.3.4": true,
-			},
-			FreeIP:       "",
-			ErrorComment: "OnlyOneUsedIPTest: Exactly one IP is available used.",
-		},
-	}
-
-	for _, testCase := range testCaseList {
-		freeIP := getUnusedIPAddressInRange(testCase.StartIPAddress, testCase.EndIPAddress,
-			testCase.UsedIPAddressMap)
-		assert.Equal(t, freeIP, testCase.FreeIP, testCase.ErrorComment)
-	}
-
-	return
-}
-
 func TestGetCursor(t *testing.T) {
 
 	type TestCase struct {
